js: console.log stdin

+++ rust/space/addressandpointer/src/main.rs
@@ -1,5 +1,7 @@
+use num_bigint::BigInt;
+
 fn main() {
-    let a: Vec<i32> = vec![];
+    let a: Vec<BigInt> = vec![];
     let ptra = &a;
     println!("{:p}\n{:p}", &a, &ptra);
 }
